aca-camp-program-scraper/requirements.txt
pika
pymongo
python-dotenv
loads
requests
lxml
python-dateutil